Initial commit of US Ignite competition single-page site
This is a single-page Jekyll site intended for use with GitHub Pages
or another Jekyll host, like CloudCannon.

The page is primarily styled with [Tachyons] classes, which have been
customized through [tachyons-custom] with PostCSS variables. The
browser-ready tachyons.min.css is built via tachyons-cli and must be
committed to the repo. A minimal number of additional custom styles
are defined in css/main.scss, using Jekyll's built-in Sass support.

Most of the content is statically defined (with CloudCannon-
compatible "editable" classes where appropriate), but the information
about judges and mentors is managed through two corresponding Jekyll
Collections in _judges and _mentors.

[Tachyons]: http://tachyons.io
[tachyons-custom]: https://github.com/tachyons-css/tachyons-custom